Separate the DB into its own crate
