+++ backend/Cargo.toml
@@ -3,7 +3,6 @@ name = "backend"
 publish = false
 version = "0.1.0"
 edition = "2021"
-build = "build.rs"
 
 [dependencies]
 anyhow = "1.0.75"
@@ -19,6 +18,7 @@ once_cell = "1.18.0"
 regex = "1.9.5"
 reqwest = { version = "0.11.20", default-features = false, features = ["json", "rustls-tls"] }
 retronomicon-dto = { path = "../retronomicon-dto", features = ["rocket"] }
+retronomicon-db = { path = "../retronomicon-db" }
 rocket = { version = "0.5.0", features = ["json", "secrets", "tls"] }
 rocket-multipart-form-data = "0.10.6"
 rocket_db_pools = { version = "0.1.0", features = ["diesel_postgres"] }
