+++ backend/src/guards/users.rs
@@ -1,19 +1,14 @@
-use crate::db::Db;
-use crate::models;
-use crate::schema;
-use crate::schema::users;
-use diesel::prelude::*;
 use jsonwebtoken::DecodingKey;
+use retronomicon_db::models;
+use retronomicon_db::models::{User, UserTeam};
+use retronomicon_db::Db;
 use retronomicon_dto as dto;
 use rocket::http::{Cookie, CookieJar, Status};
 use rocket::outcome::{IntoOutcome, Outcome};
 use rocket::{request, Request};
-use rocket_db_pools::diesel::{AsyncConnection, RunQueryDsl};
 use rocket_okapi::OpenApiFromRequest;
-use scoped_futures::ScopedFutureExt;
 use serde::{Deserialize, Serialize};
-use serde_json::{json, Value};
-use std::collections::BTreeMap;
+use serde_json::json;
 use std::ops::{Deref, DerefMut};
 use tracing::warn;
 
@@ -32,11 +27,13 @@ impl<'r> request::FromRequest<'r> for RootUserGuard {
             .guard::<Db>()
             .await
             .expect("database connection guard");
-        let user = request.guard::<UserGuard>().await.expect("user guard");
-        schema::user_teams::table
-            .filter(schema::user_teams::user_id.eq(user.id))
-            .filter(schema::user_teams::team_id.eq(1))
-            .first::<models::UserTeam>(&mut db)
+        let user = match request.guard::<UserGuard>().await {
+            Outcome::Success(user) => user,
+            Outcome::Forward(_) => return Outcome::Forward(Status::Unauthorized),
+            Outcome::Error((status, _)) => return Outcome::Error((status, "Unauthorized".into())),
+        };
+
+        UserTeam::user_is_in_team(&mut db, user.id, 1)
             .await
             .map_err(|e| e.to_string())
             .or_error(Status::Unauthorized)
@@ -214,37 +211,16 @@ impl UserGuard {
         self.exp = expiry;
     }
 
-    pub async fn into_model(self, db: &mut Db) -> Result<models::User, diesel::result::Error> {
-        use users::dsl;
-        dsl::users
-            .filter(dsl::id.eq(self.id))
-            .first::<models::User>(db)
-            .await
+    pub async fn into_model(self, db: &mut Db) -> Result<User, diesel::result::Error> {
+        User::from_id(db, self.id).await
     }
 
-    pub fn from_model(user: models::User) -> Self {
+    pub fn from_model(user: User) -> Self {
         Self::new_unchecked(user.id, user.username, default_expiration_())
     }
 
     pub async fn from_db(db: &mut Db, id: i32) -> Result<Self, diesel::result::Error> {
-        use users::dsl;
-        dsl::users
-            .filter(dsl::id.eq(id))
-            .first::<models::User>(db)
-            .await
-            .map(Self::from_model)
-    }
-
-    pub async fn from_db_by_email(
-        db: &mut Db,
-        email: String,
-    ) -> Result<Self, diesel::result::Error> {
-        use users::dsl;
-        dsl::users
-            .filter(dsl::email.eq(email))
-            .first::<models::User>(db)
-            .await
-            .map(Self::from_model)
+        User::from_id(db, id).await.map(Self::from_model)
     }
 
     /// Create a new user or select an existing one. This should only be used
@@ -256,49 +232,37 @@ impl UserGuard {
         auth_provider: String,
         avatar_url: Option<String>,
     ) -> Result<(bool, models::User, Self), diesel::result::Error> {
-        use users::dsl;
-        db.transaction(|db| {
-            async move {
-                // Set username to None if it doesn't validate.
-                let username = username.as_deref().and_then(|u| {
-                    dto::user::Username::new(u)
-                        .map(|u| u.into_inner())
-                        .map_err(|e| {
-                            warn!("login_from_auth: invalid username: {}", e);
-                            e
-                        })
-                        .ok()
-                });
+        // Set username to None if it doesn't validate.
+        let username = username.as_deref().and_then(|u| {
+            dto::user::Username::new(u)
+                .map(|u| u.into_inner())
+                .map_err(|e| {
+                    warn!("login_from_auth: invalid username: {}", e);
+                    e
+                })
+                .ok()
+        });
 
-                let maybe_user = dsl::users
-                    .filter(dsl::email.eq(email))
-                    .filter(dsl::auth_provider.eq(&auth_provider))
-                    .first::<models::User>(db)
-                    .await
-                    .optional()?;
+        let maybe_user = User::from_auth(db, email, &auth_provider).await?;
 
-                if let Some(u) = maybe_user {
-                    return Ok((false, u.clone(), Self::from_model(u)));
-                }
+        if let Some(u) = maybe_user {
+            return Ok((false, u.clone(), Self::from_model(u)));
+        }
 
-                let user = models::User::create(
-                    db,
-                    username.as_deref(),
-                    None,
-                    avatar_url.as_deref(),
-                    email,
-                    Some(&auth_provider),
-                    None,
-                    json!({}),
-                    json!({}),
-                )
-                .await?;
+        let user = models::User::create(
+            db,
+            username.as_deref(),
+            None,
+            avatar_url.as_deref(),
+            email,
+            Some(&auth_provider),
+            None,
+            json!({}),
+            json!({}),
+        )
+        .await?;
 
-                Ok((true, user.clone(), Self::from_model(user)))
-            }
-            .scope_boxed()
-        })
-        .await
+        Ok((true, user.clone(), Self::from_model(user)))
     }
 
     pub async fn update(
@@ -306,61 +270,8 @@ impl UserGuard {
         db: &mut Db,
         form: dto::user::UserUpdate<'_>,
     ) -> Result<(), diesel::result::Error> {
-        #[derive(AsChangeset)]
-        #[diesel(table_name = schema::users)]
-        struct UserSignupChangeset<'a> {
-            username: Option<&'a str>,
-            display_name: Option<&'a str>,
-            description: Option<&'a str>,
-            links: Option<Value>,
-            metadata: Option<Value>,
-        }
-
-        db.transaction(|db| {
-            async move {
-                let mut changeset = UserSignupChangeset {
-                    username: form.username,
-                    display_name: form.display_name,
-                    description: form.description,
-                    links: None,
-                    metadata: None,
-                };
-
-                if let Some(links) = form.links.as_ref() {
-                    changeset.links = Some(serde_json::to_value(links).unwrap());
-                } else if form.add_links.is_some() || form.remove_links.is_some() {
-                    let mut links = BTreeMap::new();
-                    let user: models::User = schema::users::table.find(self.id).first(db).await?;
-
-                    if let Value::Object(user_links) = user.links {
-                        links.extend(user_links.into_iter());
-                    }
-
-                    if let Some(user_links) = form.add_links {
-                        for (k, v) in user_links.into_iter() {
-                            links.insert(k.to_string(), v.into());
-                        }
-                    }
-                    if let Some(user_links) = form.remove_links {
-                        for k in user_links.into_iter() {
-                            links.remove(&k.to_string());
-                        }
-                    }
-
-                    changeset.links = Some(serde_json::to_value(links).unwrap());
-                }
-
-                diesel::update(schema::users::table)
-                    .filter(schema::users::id.eq(self.id))
-                    .set(changeset)
-                    .execute(db)
-                    .await?;
-
-                Ok(())
-            }
-            .scope_boxed()
-        })
-        .await
+        let user = self.clone().into_model(db).await?;
+        user.update(db, form).await
     }
 
     pub fn update_cookie(&self, cookies: &CookieJar<'_>) {
