+++ backend/src/routes/v1/cores/releases.rs
@@ -1,7 +1,8 @@
-use crate::db::Db;
-use crate::types::FetchModel;
+use crate::guards;
 use crate::utils::acls;
-use crate::{guards, models};
+use retronomicon_db::models;
+use retronomicon_db::types::FetchModel;
+use retronomicon_db::Db;
 use retronomicon_dto as dto;
 use rocket::data::ToByteUnit;
 use rocket::http::{ContentType, Header, Status};
