+++ backend/src/routes/v1/games.rs
@@ -1,5 +1,6 @@
-use crate::db::Db;
-use crate::{guards, models};
+use crate::guards;
+use retronomicon_db::models;
+use retronomicon_db::Db;
 use retronomicon_dto as dto;
 use rocket::http::Status;
 use rocket::serde::json::Json;
