+++ backend/src/routes/v1/platforms.rs
@@ -1,6 +1,7 @@
-use crate::db::Db;
-use crate::types::FetchModel;
-use crate::{guards, models};
+use crate::guards;
+use retronomicon_db::models;
+use retronomicon_db::types::FetchModel;
+use retronomicon_db::Db;
 use retronomicon_dto as dto;
 use rocket::http::Status;
 use rocket::serde::json::Json;
