+++ retronomicon-db/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "retronomicon-db"
+version = "0.1.0"
+edition = "2021"
+build = "build.rs"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+chrono = { version = "0.4.26", features = ["serde"] }
+diesel = { version = "2", features = ["postgres", "chrono", "extras"] }
+diesel_migrations = { version = "2.1.0", features = ["postgres"] }
+jsonwebtoken = "8.1.1"
+md5 = "0.7.0"
+retronomicon-dto = { path = "../retronomicon-dto" }
+rocket = { version = "0.5.0", features = ["json", "secrets", "tls"] }
+rocket_db_pools = { version = "0.1.0", features = ["diesel_postgres"] }
+rustls = "0.21.7"
+rustls-native-certs = "0.6.2"
+rustls-pemfile = "2.0.0-alpha.1"
+scoped-futures = "0.1.3"
+serde = { version = "1.0.186", features = ["derive"] }
+serde_json = "1.0.64"
+sha1 = "0.10.6"
+sha2 = "0.10.8"
+tokio-postgres = "0.7.10"
+tokio-postgres-rustls = "0.10.0"
+tracing = "0.1.37"
