+++ retronomicon-db/src/models/systems.rs
@@ -1,5 +1,5 @@
-use crate::db::Db;
 use crate::types::FromIdOrSlug;
+use crate::Db;
 use crate::{models, schema};
 use diesel::prelude::*;
 use diesel::{Identifiable, Queryable};
