+++ rust-toolchain.toml
@@ -1,0 +1,3 @@
+[toolchain]
+channel = "1.75"
+components = [ "clippy", "rustfmt", "rust-docs", "rust-std" ]
